+++ src/components/community/CommunityTab.tsx
@@ -3,10 +3,11 @@ import { ReactComponent as All } from "@assets/svg/all.svg";
 import { ReactComponent as Communication } from "@assets/svg/comment.svg";
 import { ReactComponent as Promotion } from "@assets/svg/invite.svg";
 import { ReactComponent as Review } from "@assets/svg/review.svg";
-import { ITabs } from "@/types/community";
+
 import { useRecoilState, useSetRecoilState } from "recoil";
 import { communityTabState } from "@/state/communityTabState";
 import { searchCriteria } from "@/state/searchCriteria";
+import { ITabs } from "@/interface/community";
 
 const TABS: ITabs[] = [
 	{
