+++ src/components/communityDetail/CommunityLikeAndComment.tsx
@@ -1,5 +1,5 @@
 import useAxios from "@/hooks/useAxios";
-import { ICommunityList } from "@/interface/community";
+import { ICommunityItem } from "@/interface/community";
 import { loginState } from "@/state/loginState";
 import { alertHandler } from "@/utils/alert";
 import { ReactComponent as Comment } from "@assets/svg/comment.svg";
@@ -11,7 +11,7 @@ import { useRecoilValue } from "recoil";
 import Swal from "sweetalert2";
 
 interface IProps {
-	readonly communityInfo: ICommunityList;
+	readonly communityInfo: ICommunityItem;
 }
 
 const CommunityLikeAndComment = ({ communityInfo }: IProps) => {
